+++ app/src/main/res/layout/activity_main.xml
@@ -12,7 +12,7 @@
         android:layout_height="match_parent"
         android:foreground="@drawable/background_gradient" />
 
-    <com.infotsav.test.MainMenuUtils.ItemsCountView
+    <com.infotsav.test.CardViewLib.ItemsCountView
         android:id="@+id/items_count_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
