+++ app/src/main/res/layout/activity_welcome.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.infotsav.test.WelcomeActivity"
+    tools:context="com.infotsav.test.Main_Activities.WelcomeActivity"
     android:orientation="vertical"
     android:gravity="center"
     android:background="@color/colorPrimaryDark">
